+++ lib/html_sigs.mli
@@ -1368,6 +1368,15 @@ module type T = sig
 
   val meta : ([< | meta_attrib], [> | meta]) nullary
 
+  val meta_itemprop :
+    itemprop: string list wrap ->
+    ([< | meta_attrib], [> | meta_with_itemprop]) nullary
+  (** [meta_itemprop ~itemprop ()] is a [meta] element annotating the
+      surrounding microdata item. Unlike {!meta}, it is phrasing content, so it
+      is allowed in the body and not only in the head.
+      @see <https://html.spec.whatwg.org/multipage/semantics.html#the-meta-element>
+      meta element specification *)
+
   (** {3 Style Sheets} *)
 
   val style :